+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Confirmation Dialog/ConfirmationDialogButton.swift
@@ -14,13 +14,13 @@ public struct ConfirmationDialogButton: ConfirmationDialogButtonProtocol, Sendab
     private var isEnabled: Bool = true
     private let role: ButtonRole?
     private let title: String
-    private let action: (() -> Void)?
+    private let action: (@Sendable () -> Void)?
     
     // MARK: Initializers
     /// Initializes `ConfirmationDialogButton`.
     public init(
         role: ButtonRole? = nil,
-        action: (() -> Void)?,
+        action: (@Sendable () -> Void)?,
         title: String
     ) {
         self.role = role