+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### 7.1.0(104)
+### [7.1.0(104)](https://github.com/VakhoKontridze/VCore/releases/tag/7.1.0) — *2024 10 24*
 
 General